// max-mini-app/src/api/models/Organization.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MaxQueue Orchestrator API
 * API, which is the entry point to the MaxQueue project
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserRole & string } from './UserRole &amp; string';
import {
    UserRole &amp; stringFromJSON,
    UserRole &amp; stringFromJSONTyped,
    UserRole &amp; stringToJSON,
    UserRole &amp; stringToJSONTyped,
} from './UserRole &amp; string';

/**
 * 
 * @export
 * @interface Organization
 */
export interface Organization {
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Organization
     */
    name?: string;
    /**
     * 
     * @type {UserRole & string}
     * @memberof Organization
     */
    role?: UserRole & string;
    /**
     * 
     * @type {number}
     * @memberof Organization
     */
    amountOfQueues?: number;
}

/**
 * Check if a given object implements the Organization interface.
 */
export function instanceOfOrganization(value: object): value is Organization {
    return true;
}

export function OrganizationFromJSON(json: any): Organization {
    return OrganizationFromJSONTyped(json, false);
}

export function OrganizationFromJSONTyped(json: any, ignoreDiscriminator: boolean): Organization {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'role': json['role'] == null ? undefined : UserRole &amp; stringFromJSON(json['role']),
        'amountOfQueues': json['amountOfQueues'] == null ? undefined : json['amountOfQueues'],
    };
}

export function OrganizationToJSON(json: any): Organization {
    return OrganizationToJSONTyped(json, false);
}

export function OrganizationToJSONTyped(value?: Organization | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'role': UserRole &amp; stringToJSON(value['role']),
        'amountOfQueues': value['amountOfQueues'],
    };
}