// max-mini-app/src/api/models/QueueEntryInUserResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MaxQueue Orchestrator API
 * API, which is the entry point to the MaxQueue project
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { QueueEntryStatus } from './QueueEntryStatus';
import {
    QueueEntryStatusFromJSON,
    QueueEntryStatusFromJSONTyped,
    QueueEntryStatusToJSON,
    QueueEntryStatusToJSONTyped,
} from './QueueEntryStatus';

/**
 * 
 * @export
 * @interface QueueEntryInUserResponse
 */
export interface QueueEntryInUserResponse {
    /**
     * 
     * @type {string}
     * @memberof QueueEntryInUserResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof QueueEntryInUserResponse
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof QueueEntryInUserResponse
     */
    peopleInFront?: number;
    /**
     * 
     * @type {QueueEntryStatus}
     * @memberof QueueEntryInUserResponse
     */
    status?: QueueEntryStatus;
}



/**
 * Check if a given object implements the QueueEntryInUserResponse interface.
 */
export function instanceOfQueueEntryInUserResponse(value: object): value is QueueEntryInUserResponse {
    return true;
}

export function QueueEntryInUserResponseFromJSON(json: any): QueueEntryInUserResponse {
    return QueueEntryInUserResponseFromJSONTyped(json, false);
}

export function QueueEntryInUserResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): QueueEntryInUserResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'peopleInFront': json['peopleInFront'] == null ? undefined : json['peopleInFront'],
        'status': json['status'] == null ? undefined : QueueEntryStatusFromJSON(json['status']),
    };
}

export function QueueEntryInUserResponseToJSON(json: any): QueueEntryInUserResponse {
    return QueueEntryInUserResponseToJSONTyped(json, false);
}

export function QueueEntryInUserResponseToJSONTyped(value?: QueueEntryInUserResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'peopleInFront': value['peopleInFront'],
        'status': QueueEntryStatusToJSON(value['status']),
    };
}